% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanapiwrapper.R
\name{accessRestrictions}
\alias{accessRestrictions}
\title{Get access restrictions for a specific Santiment metric or query.}
\usage{
accessRestrictions(name)
}
\arguments{
\item{name}{name of Santiment metric or query}
}
\value{
single row with columns (isRestricted, name, restrictedFrom, restrictedTo, type)
}
\description{
Get access restrictions for a specific Santiment metric or query.
}
\examples{
accessRestrictions('daily_active_addresses')

}
