% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanapiwrapper.R
\name{availableMetrics}
\alias{availableMetrics}
\title{Get the available Santiment metrics for a specific project.}
\usage{
availableMetrics(slug)
}
\arguments{
\item{slug}{project}
}
\value{
vector of available metrics
}
\description{
Returns all availabale timeseries metrics (note not the histogram metrics) for a slug.
}
\examples{
availableMetrics('ethereum')

}
