% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanapiwrapper.R
\name{availableSince}
\alias{availableSince}
\title{Get the earliest date for which the Santiment metric is available.}
\usage{
availableSince(metric, slug)
}
\arguments{
\item{metric}{metric}

\item{slug}{project}
}
\value{
earliest date
}
\description{
Get the earliest date for which the Santiment metric is available.
}
\examples{
availableSince('daily_active_addresses', 'ethereum')

}
