% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanapiwrapper.R
\name{availableSlugs}
\alias{availableSlugs}
\title{Get the available slugs for a specific metric}
\usage{
availableSlugs(metric)
}
\arguments{
\item{metric}{Santiment metric}
}
\value{
vector of available slugs
}
\description{
Get the available slugs for a specific metric
}
\examples{
availableSlugs('daily_active_addresses')

}
