% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanapiwrapper.R
\name{santimentMetric}
\alias{santimentMetric}
\title{Get a metric from the Santiment GraphQL API.}
\usage{
santimentMetric(
  metric,
  slug,
  from = "2019-01-01",
  to = "2020-01-01",
  aggregation = "SUM",
  selector_option
)
}
\arguments{
\item{metric}{Santiment metric}

\item{slug}{project (e.g. ethereum, bitcoin, tezos ...)}

\item{from}{start of the time window}

\item{to}{end of the time window}

\item{aggregation}{aggregation (SUM, AVG, MEDIAN, MIN, MAX, FIRST, LAST)}

\item{selector_option}{optional string for selector field, for example "holdersCount:10" for "amount_in_top_holders" metric, see "selector: {slug: "ethereum", holdersCount:10}"}
}
\value{
table with columns (date, <metric>)
}
\description{
Checks if the metric is available for the slug. If not, a constant vector is returned.
Checks if the metric is available for this time window. For unavailable dates, zeros are returned.
}
\details{
https://neuro.santiment.net
https://graphql.org
}
\examples{
santimentMetric('dev_activity', 'ethereum')

}
