% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanapiwrapper.R
\name{setSantimentVariables}
\alias{setSantimentVariables}
\title{Set API key and API URL for Santiment GraphQL queries.}
\usage{
setSantimentVariables(key, url = "https://api.santiment.net/graphql")
}
\arguments{
\item{key}{API key}

\item{url}{GraphQL server URL}
}
\description{
Note that setting API key and API URL is optional. Without them queries will default to a free Santiment account.
}
\examples{
setSantimentVariables("y4i3n5xk6p")

}
