% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanapiwrapper.R
\name{santimentQuery}
\alias{santimentQuery}
\title{Execute Santiment query.}
\usage{
santimentQuery(query_string)
}
\arguments{
\item{query_string}{query string}
}
\value{
result
}
\description{
A new GraphQL client is generated. A new query is generated, filled with the query string and then executed.
}
\examples{
santimentQuery('{projectBySlug(slug: "santiment") {availableQueries}}')

}
